🌱 Split settings/sections GitHubTokenSection.tsx and LocalClustersSection.tsx
Fixes #21797

Extract useGitHubToken hook (token validation/scope-check/save state)
from GitHubTokenSection.tsx, and useLocalClusters hook +
LocalClustersSection.parts (VClusterSection, KubeVirtSection) from
LocalClustersSection.tsx. Token handling and cluster dedup behaviour
unchanged; both files now under 350 lines with fewer hook calls.
